# spec/api_contracts.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Virtual Try-On API",
    "description": "Comprehensive API specification for the Virtual Try-On application supporting avatar scanning, 3D product try-on, e-commerce, and user management",
    "version": "1.0.0",
    "contact": {
      "name": "API Support",
      "email": "api-support@tryon.com",
      "url": "https://docs.tryon.com/support"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.tryon.com/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.tryon.com/v1",
      "description": "Staging server"
    },
    {
      "url": "https://dev-api.tryon.com/v1",
      "description": "Development server"
    }
  ],
...
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": ["authentication"],
        "summary": "User login",
        "description": "Authenticate user with email/password and return access token",
        "operationId": "userLogin",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              },
              "examples": {
                "basic_login": {
                  "summary": "Basic email/password login",
                  "value": {
                    "email": "user@example.com",
                    "password": "securePassword123"
                  }
                },
                "social_login": {
                  "summary": "Social login with provider",
                  "value": {
                    "provider": "google",
                    "providerToken": "eyJhbGciOiJIUzI1NiIs..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "headers": {
              "Set-Cookie": {
                "description": "HTTP-only refresh token cookie",
                "schema": {
                  "type": "string",
                  "example": "refresh_token=abc123; HttpOnly; Secure; SameSite=Strict; Max-Age=604800"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                },
                "examples": {
                  "success": {
                    "value": {
                      "success": true,
                      "data": {
                        "user": {
                          "id": "usr_123456789",
                          "email": "user@example.com",
                          "firstName": "John",
                          "lastName": "Doe",
                          "avatar": "https://cdn.tryon.com/avatars/usr_123456789.jpg",
                          "emailVerified": true,
                          "createdAt": "2025-01-15T10:30:00Z"
                        },
                        "tokens": {
                          "accessToken": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                          "accessTokenExpires": "2025-01-15T11:30:00Z",
                          "refreshToken": "refresh_token_abc123",
                          "refreshTokenExpires": "2025-01-22T10:30:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "success": false,
                  "error": {
                    "code": "INVALID_CREDENTIALS",
                    "message": "Invalid email or password",
                    "details": {
                      "field": "email",
                      "reason": "Email or password is incorrect"
                    }
                  }
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RateLimitError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": ["authentication"],
        "summary": "User registration",
        "description": "Register a new user account with email verification",
        "operationId": "userRegistration",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              },
              "examples": {
                "basic_register": {
                  "value": {
                    "email": "newuser@example.com",
                    "password": "SecurePass123!",
                    "firstName": "Jane",
                    "lastName": "Smith",
                    "dateOfBirth": "1995-08-15",
                    "agreeToTerms": true,
                    "marketingConsent": false
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "409": {
            "description": "Email already exists",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": ["authentication"],
        "summary": "Refresh access token",
        "description": "Refresh expired access token using refresh token",
        "operationId": "refreshToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token refreshed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshTokenResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid or expired refresh token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/profile": {
      "get": {
        "tags": ["profile"],
        "summary": "Get user profile",
        "description": "Retrieve the authenticated user's profile information",
        "operationId": "getUserProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["profile"],
        "summary": "Update user profile",
        "description": "Update the authenticated user's profile information",
        "operationId": "updateUserProfile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid profile data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token obtained from login endpoint"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for service-to-service authentication"
      }
    },
    "schemas": {
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "User's password",
            "minLength": 8,
            "example": "securePassword123"
          },
          "provider": {
            "type": "string",
            "enum": ["google", "apple", "facebook"],
            "description": "OAuth provider for social login"
          },
          "providerToken": {
            "type": "string",
            "description": "OAuth access token from provider"
          },
          "rememberMe": {
            "type": "boolean",
            "description": "Extend session duration",
            "default": false
          }
        }
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Request success status"
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "tokens": {
                "$ref": "#/components/schemas/AuthTokens"
              }
            }
          }
        }
      },
      "RegisterRequest": {
        "type": "object",
        "required": ["email", "password", "firstName", "lastName", "agreeToTerms"],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "description": "User's email address"
          },
          "password": {
            "type": "string",
            "format": "password",
            "description": "User's password",
            "minLength": 8,
            "pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,}$"
          },
          "firstName": {
            "type": "string",
            "description": "User's first name",
            "minLength": 1,
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "description": "User's last name",
            "minLength": 1,
            "maxLength": 50
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date",
            "description": "User's date of birth (YYYY-MM-DD)",
            "minimum": "1920-01-01",
            "maximum": "2010-12-31"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female", "non_binary", "prefer_not_to_say"],
            "description": "User's gender"
          },
          "phoneNumber": {
            "type": "string",
            "pattern": "^\\+?[1-9]\\d{1,14}$",
            "description": "User's phone number in international format"
          },
          "agreeToTerms": {
            "type": "boolean",
            "description": "User agreement to terms and conditions",
            "enum": [true]
          },
          "marketingConsent": {
            "type": "boolean",
            "description": "User consent to marketing communications",
            "default": false
          }
        }
      },
      "RegisterResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "user": {
                "$ref": "#/components/schemas/User"
              },
              "verificationEmailSent": {
                "type": "boolean",
                "description": "Whether verification email was sent"
              },
              "emailVerificationRequired": {
                "type": "boolean",
                "description": "Email verification is required before login"
              }
            }
          }
        }
      },
      "RefreshTokenRequest": {
        "type": "object",
        "required": ["refreshToken"],
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Valid refresh token"
          }
        }
      },
      "RefreshTokenResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "properties": {
              "tokens": {
                "$ref": "#/components/schemas/AuthTokens"
              }
            }
          }
        }
      },
...
      "UserProfileResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "$ref": "#/components/schemas/UserProfile"
          }
        }
      },
      "UpdateProfileRequest": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "lastName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "gender": {
            "type": "string",
            "enum": ["male", "female", "non_binary", "prefer_not_to_say"]
          },
          "phoneNumber": {
            "type": "string",
            "pattern": "^\\+?[1-9]\\d{1,14}$"
          },
          "preferences": {
            "type": "object",
            "properties": {
              "notifications": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "boolean",
                    "default": true
                  },
                  "push": {
                    "type": "boolean",
                    "default": true
                  },
                  "marketing": {
                    "type": "boolean",
                    "default": false
                  }
                }
              },
              "privacy": {
                "type": "object",
                "properties": {
                  "profileVisibility": {
                    "type": "string",
                    "enum": ["public", "friends", "private"],
                    "default": "private"
                  },
                  "dataSharing": {
                    "type": "boolean",
                    "default": false
                  }
                }
              }
            }
          },
          "addresses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Address"
            }
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "description": "Error code for client handling"
              },
              "message": {
                "type": "string",
                "description": "Human-readable error message"
              },
              "details": {
                "type": "object",
                "description": "Additional error details"
              }
            }
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["VALIDATION_ERROR"]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "properties": {
                  "fields": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "field": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "code": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "RateLimitError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [false]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": ["RATE_LIMIT_EXCEEDED"]
              },
              "message": {
                "type": "string"
              },
              "details": {
                "type": "object",
                "properties": {
                  "limit": {
                    "type": "integer"
                  },
                  "remaining": {
                    "type": "integer"
                  },
                  "resetTime": {
                    "type": "string",
                    "format": "date-time"
                  }
                }
              }
            }
          }
        }
      },
...
      "User": {
        "type": "object",
        "required": ["id", "email", "firstName", "lastName"],
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique user identifier"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "format": "uri"
          },
          "emailVerified": {
            "type": "boolean"
          },
          "phoneVerified": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}